Fix type errors and rename LinkButton file
Clean up duplication in import path.
