+++ frontend/src/components/Header/Header.tsx
@@ -7,7 +7,7 @@ import {
 import ProjectPicker from '@components/Header/components/ProjectPicker/ProjectPicker'
 import Notifications from '@components/Header/Notifications/NotificationsV2'
 import { linkStyle } from '@components/Header/styles.css'
-import { LinkButton } from '@components/LinkButton/LinkButton'
+import { LinkButton } from '@components/LinkButton'
 import { useGetBillingDetailsForProjectQuery } from '@graph/hooks'
 import { Maybe, PlanType, Project } from '@graph/schemas'
 import {
@@ -156,6 +156,7 @@ export const Header = () => {
 												}
 												to={`/${project_id}/${p.key}`}
 												key={p.key}
+												trackingId={`header-link-click-${p.key}`}
 											>
 												{titleCaseString(p.key)}
 											</LinkButton>
